+++ src/careamics/lightning/dataset_ng/callbacks/prediction_writer/__init__.py
@@ -1,7 +1,7 @@
 """A package for the `PredictionWriterCallback` class and utilities."""
 
 __all__ = [
-    "CacheTiles",
+    "CachedTiles",
     "PredictionWriterCallback",
     "WriteImage",
     "WriteStrategy",
@@ -13,12 +13,13 @@ __all__ = [
     "select_write_func",
 ]
 
+from .cached_tiles_strategy import CachedTiles
 from .file_path_utils import create_write_file_path
 from .prediction_writer_callback import (
     PredictionWriterCallback,
     decollate_image_region_data,
 )
-from .write_strategy import CacheTiles, WriteImage, WriteStrategy
+from .write_strategy import WriteImage, WriteStrategy
 from .write_strategy_factory import (
     create_write_strategy,
     select_write_extension,
